+++ spaceobject.cpp
@@ -1,6 +1,6 @@
 #include "spaceobject.h"
 
-SpaceObject::SpaceObject(const QVector2D& position, int resources, const QColor &color, QObject *parent) :
+SpaceObject::SpaceObject(const QVector2D& position, qreal resources, const QColor &color, QObject *parent) :
     Drawable(parent)
 {
     m_position = position;
